rename pulse_shape to blob_shape in all files
