+++ blobmodel/__init__.py
@@ -3,4 +3,4 @@ from .blobs import Blob
 from .plotting import show_model
 from .stochasticality import BlobFactory, DefaultBlobFactory
 from .geometry import Geometry
-from .pulse_shape import AbstractBlobShape, BlobShapeImpl
+from .blob_shape import AbstractBlobShape, BlobShapeImpl
