+++ blobmodel/blobs.py
@@ -4,7 +4,7 @@ import warnings
 from typing import Tuple, Union
 from nptyping import NDArray
 import numpy as np
-from .pulse_shape import AbstractBlobShape
+from .blob_shape import AbstractBlobShape
 import cmath
